move Orchestrator class to a separate directory and update imports across the codebase
