+++ tests/orchestrator/conftest.py
@@ -1,4 +1,4 @@
-from unittest.mock import MagicMock, patch
+from unittest.mock import MagicMock
 
 import pytest
 from sqlmodel import Session, SQLModel
@@ -13,7 +13,6 @@ from geoparser.db.models import (
     Toponym,
     ToponymCreate,
 )
-from geoparser.geoparserv2.geoparserv2 import GeoparserV2
 from geoparser.modules.interfaces import (
     AbstractRecognitionModule,
     AbstractResolutionModule,
@@ -41,26 +40,6 @@ def test_project(test_db: Session):
     test_db.commit()
     test_db.refresh(project)
     return project
-
-
-@pytest.fixture
-def mock_get_db(test_db):
-    """Mock the get_db function to return our test database session."""
-    with patch("geoparser.geoparserv2.geoparserv2.get_db") as mock:
-        mock.return_value = iter([test_db])
-        yield mock
-
-
-@pytest.fixture
-def geoparser_with_existing_project(mock_get_db, test_project):
-    """Create a GeoparserV2 instance with an existing project."""
-    return GeoparserV2(project_name=test_project.name)
-
-
-@pytest.fixture
-def geoparser_with_new_project(mock_get_db):
-    """Create a GeoparserV2 instance with a new project."""
-    return GeoparserV2(project_name="new-test-project")
 
 
 @pytest.fixture
